removes unneeded cast

+++ src/tensorflow_time_series_dataset/preprocessors/groupby_dataset_generator.py
@@ -19,7 +19,7 @@ class GroupbyDatasetGenerator:
 
         def generator():
             for _, d in grpd:
-                yield tf.cast(d[self.columns].values, self.dtype)
+                yield d[self.columns].values
 
         return generator
